% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generic Non-inferiority Function.R
\name{NIPLOT}
\alias{NIPLOT}
\title{NIPlot}
\usage{
NIPLOT(mean, lbound, ubound, title = "Non-Inferiority Plot",
  subtitle = expression(paste(mu, "1 - ", mu, "2")), delta = 1, limit = 3,
  filename = "niplot.png", direction = "positive")
}
\arguments{
\item{mean}{sample mean}

\item{lbound}{CI lower bound}

\item{ubound}{CI upper bound}

\item{title}{Optional title}

\item{subtitle}{optional subtitle}

\item{delta}{non-inferiority margin}

\item{filename}{optional filename - saves to current directory}
}
\description{
This function allows you to make non-inferiority plots.
}
\examples{
NIPLOT()
}
\keyword{non-inferiority}
